requirements.txt
openai>=1.30.1
python-telegram-bot>=20.3
python-dotenv>=1.0.0
tqdm>=4.64.0
